moved debug heap
